fixing routes + page sidebar

+++ routes/web.php
@@ -3,8 +3,8 @@
 use Illuminate\Support\Facades\Route;
 
 use App\Http\Controllers\AuthController;
-use App\Http\Controllers\OperatorAdminController;
 use App\Http\Controllers\AdminController;
+use App\Http\Controllers\OperatorAdminController;
 use App\Http\Controllers\DriverAdminController;
 
 
@@ -18,8 +18,6 @@ use App\Http\Controllers\DriverAdminController;
 | be assigned to the "web" middleware group. Make something great!
 |
 */
-Route::get('main', [AdminController::class, 'main'])->name('main');
-Route::get('dashboard/admin', [AdminController::class, 'dashboard'])->name('dashboard.admin');
 
 // Auth
 Route::get('login', [AuthController::class, 'showLoginForm'])->name('login');
@@ -30,25 +28,26 @@ Route::post('forgot-password', [AuthController::class, 'sendResetLink'])->name('
 
 // Admin routes
 Route::middleware(['auth', 'role:admin'])->prefix('admin')->group(function () {
+    Route::get('dashboard', [AdminController::class, 'dashboard'])->name('dashboard');
     // Operator CRUD
-    Route::get('operators', [OperatorAdminController::class, 'index'])->name('operator.index');
-    Route::get('operators/create', [OperatorAdminController::class, 'create'])->name('operator.create');
-    Route::get('operators/detail', [OperatorAdminController::class, 'detail'])->name('operator.detail');
-    Route::post('operators', [OperatorAdminController::class, 'store'])->name('operator.store');
-    Route::get('operators/{id}/edit', [OperatorAdminController::class, 'edit'])->name('operator.edit');
-    Route::get('operators/edit', [OperatorAdminController::class, 'edit'])->name('operator.edit');
-    Route::put('operators/{id}', [OperatorAdminController::class, 'update'])->name('operator.update');
-    Route::delete('operators/{id}', [OperatorAdminController::class, 'destroy'])->name('operator.destroy');
+    Route::get('operators', [OperatorAdminController::class, 'index'])->name('admin.operator.index');
+    Route::get('operators/create', [OperatorAdminController::class, 'create'])->name('admin.operator.create');
+    Route::get('operators/detail', [OperatorAdminController::class, 'detail'])->name('admin.operator.detail');
+    Route::post('operators', [OperatorAdminController::class, 'store'])->name('admin.operator.store');
+    Route::get('operators/{id}/edit', [OperatorAdminController::class, 'edit'])->name('admin.operator.edit');
+    Route::get('operators/edit', [OperatorAdminController::class, 'edit'])->name('admin.operator.edit');
+    Route::put('operators/{id}', [OperatorAdminController::class, 'update'])->name('admin.operator.update');
+    Route::delete('operators/{id}', [OperatorAdminController::class, 'destroy'])->name('admin.operator.destroy');
 
     // Driver CRUD
-    Route::get('drivers', [DriverAdminController::class, 'index'])->name('driver.index');
-    Route::get('drivers/create', [DriverAdminController::class, 'create'])->name('driver.create');
-    Route::get('drivers/detail', [DriverAdminController::class, 'detail'])->name('driver.detail');
-    Route::post('drivers', [DriverAdminController::class, 'store'])->name('driver.store');
+    Route::get('drivers', [DriverAdminController::class, 'index'])->name('admin.driver.index');
+    Route::get('drivers/create', [DriverAdminController::class, 'create'])->name('admin.driver.create');
+    Route::get('drivers/detail', [DriverAdminController::class, 'detail'])->name('admin.driver.detail');
+    Route::post('drivers', [DriverAdminController::class, 'store'])->name('admin.driver.store');
     Route::get('drivers/{id}/edit', [DriverAdminController::class, 'edit'])->name('driver.edit');
-    Route::get('drivers/edit', [DriverAdminController::class, 'edit'])->name('driver.edit');
-    Route::put('drivers/{id}', [DriverAdminController::class, 'update'])->name('driver.update');
-    Route::delete('drivers/{id}', [DriverAdminController::class, 'destroy'])->name('driver.destroy');
+    Route::get('drivers/edit', [DriverAdminController::class, 'edit'])->name('admin.driver.edit');
+    Route::put('drivers/{id}', [DriverAdminController::class, 'update'])->name('admin.driver.update');
+    Route::delete('drivers/{id}', [DriverAdminController::class, 'destroy'])->name('admin.driver.destroy');
 });
 
 // Operator routes
